chore(store/api/yt): better type of YTEndpointApiArgs
