+++ packages/ui/src/ui/store/api/yt/checkPermissions/index.ts
@@ -1,8 +1,8 @@
 import {DEFAULT_UPDATER_TIMEOUT} from '../../../../hooks/use-updater';
 import {useSelector} from '../../../../store/redux-hooks';
-import {selectCluster} from '../../../../store/selectors/global/cluster';
 import {selectUseAutoRefresh} from '../../../../store/selectors/settings/settings-ts';
 import {type OverrideDataType} from '../types';
+import {useEffectiveClusterArgs} from '../utils';
 import {ytApi} from '../ytApi';
 import {type Params, checkPermission} from './endpoint';
 
@@ -17,21 +17,14 @@ const checkPermissionApi = ytApi.injectEndpoints({
 
 export function useCheckPermissionQuery(params: Params) {
     const useAutoRefresh = useSelector(selectUseAutoRefresh);
-    const cluster = useSelector(selectCluster);
 
     const options = {
         pollingInterval: useAutoRefresh ? DEFAULT_UPDATER_TIMEOUT : undefined,
         skipPollingIfUnfocused: true,
     };
 
-    const customParams =
-        'setup' in params
-            ? params
-            : {
-                  ...params,
-                  cluster,
-              };
+    const effectiveParams = useEffectiveClusterArgs(params);
 
-    const res = checkPermissionApi.useCheckPermissionsQuery(customParams, options);
+    const res = checkPermissionApi.useCheckPermissionsQuery(effectiveParams, options);
     return res as OverrideDataType<typeof res, (typeof res)['data']>;
 }
