+++ packages/ui/src/ui/store/api/yt/get/index.ts
@@ -1,8 +1,8 @@
 import {DEFAULT_UPDATER_TIMEOUT} from '../../../../hooks/use-updater';
 import {useSelector} from '../../../../store/redux-hooks';
-import {selectCluster} from '../../../../store/selectors/global/cluster';
 import {selectUseAutoRefresh} from '../../../../store/selectors/settings/settings-ts';
 import {type OverrideDataType} from '../types';
+import {useEffectiveClusterArgs} from '../utils';
 import {ytApi} from '../ytApi';
 import {get} from './endpoint';
 
@@ -17,21 +17,14 @@ export const getApi = ytApi.injectEndpoints({
 
 export function useGetQuery<T>(args: Parameters<typeof get>[0]) {
     const useAutoRefresh = useSelector(selectUseAutoRefresh);
-    const cluster = useSelector(selectCluster);
 
     const options = {
         pollingInterval: useAutoRefresh ? DEFAULT_UPDATER_TIMEOUT : undefined,
         skipPollingIfUnfocused: true,
     };
 
-    const customArgs =
-        'setup' in args
-            ? args
-            : {
-                  ...args,
-                  cluster,
-              };
+    const effectiveArgs = useEffectiveClusterArgs(args);
 
-    const res = getApi.useGetQuery(customArgs, options);
+    const res = getApi.useGetQuery(effectiveArgs, options);
     return res as OverrideDataType<typeof res, T>;
 }
