+++ packages/ui/src/ui/store/api/yt/selectRows/index.ts
@@ -1,8 +1,8 @@
 import {DEFAULT_UPDATER_TIMEOUT} from '../../../../hooks/use-updater';
 import {useSelector} from '../../../../store/redux-hooks';
-import {selectCluster} from '../../../../store/selectors/global/cluster';
 import {selectUseAutoRefresh} from '../../../../store/selectors/settings/settings-ts';
 import {type OverrideDataType} from '../types';
+import {useEffectiveClusterArgs} from '../utils';
 import {ytApi} from '../ytApi';
 import {selectRows} from './endpoint';
 import {type SelectRowsResponse} from '../../../../../shared/yt-types';
@@ -20,21 +20,14 @@ export function useSelectRowsQuery<RowT extends Record<string, unknown>>(
     args: Parameters<typeof selectRows>[0],
 ) {
     const useAutoRefresh = useSelector(selectUseAutoRefresh);
-    const cluster = useSelector(selectCluster);
 
     const options = {
         pollingInterval: useAutoRefresh ? DEFAULT_UPDATER_TIMEOUT : undefined,
         skipPollingIfUnfocused: true,
     };
 
-    const customArgs =
-        'setup' in args
-            ? args
-            : {
-                  ...args,
-                  cluster,
-              };
+    const effectiveArgs = useEffectiveClusterArgs(args);
 
-    const res = selectRowsApi.useSelectRowsQuery(customArgs, options);
+    const res = selectRowsApi.useSelectRowsQuery(effectiveArgs, options);
     return res as OverrideDataType<typeof res, SelectRowsResponse<RowT>>;
 }
